% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czestotliwosc_bank.R
\name{czestotliwosc_bank}
\alias{czestotliwosc_bank}
\title{Funkcja wyznacza srednia czestotliwosc rzeczownikow wybranego banku}
\usage{
czestotliwosc_bank(nazwa_banku, data_start = "2013-09-01",
  data_koniec = "2014-01-01")
}
\arguments{
\item{nazwa_banku}{Nazwa banku, obecnie do wyboru "ALIOR", "ING", "BZWBK"}

\item{data_start}{Poczatek rozpatrywanego okresu}

\item{data_koniec}{Koniec rozpatrywanego okresu}
}
\value{
wektor named_num czestotliwosci danego slowa
}
\description{
Funkcja czestotliwosc_bank() wyznacza srednia czestotliwosc rzeczownikow w formie podstawoej
wystepujacych we wpisach dla zadanego banku i przedzialu czasowego
}

