% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daty_ALIOR.R
\docType{data}
\name{daty_ALIOR}
\alias{daty_ALIOR}
\title{Zbior danych daty_ALIOR}
\format{wektor dat dlugosci 6920}
\usage{
data(daty_ALIOR)
}
\description{
Zbior zawiera daty kazdego wpisu ramki wpisy_ALIOR
}
\keyword{datasets}

