% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daty_BZWBK.R
\docType{data}
\name{daty_BZWBK}
\alias{daty_BZWBK}
\title{Zbior danych daty_BZWBK}
\format{wektor dat dlugosci 25760}
\usage{
data(daty_BZWBK)
}
\description{
Zbior zawiera daty kazdego wpisu ramki wpisy_BZWBK
}
\keyword{datasets}

