% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daty_ING.R
\docType{data}
\name{daty_ING}
\alias{daty_ING}
\title{Zbior danych daty_ING}
\format{wektor dat dlugosci 6977}
\usage{
data(daty_ING)
}
\description{
Zbior zawiera daty kazdego wpisu ramki wpisy_ING
}
\keyword{datasets}

