% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nouns_ALIOR.R
\docType{data}
\name{nouns_ALIOR}
\alias{nouns_ALIOR}
\title{Zbior danych nouns_ALIoR}
\format{lista wektorow napisow dlugosci 6920}
\usage{
data(nouns_ALIOR)
}
\description{
Zbior zawiera rzeczowniki w formie podstawowej z kazdego wpisu z ramki wpisy_ALIOR
}
\keyword{datasets}

