% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nouns_BZWBK.R
\docType{data}
\name{nouns_BZWBK}
\alias{nouns_BZWBK}
\title{Zbior danych nouns_BZWBK}
\format{lista wektorow napisow dlugosci 25760}
\usage{
data(nouns_BZWBK)
}
\description{
Zbior zawiera rzeczowniki w formie podstawowej
z kazdego wpisu z ramki wpisy_BZWBK
}
\keyword{datasets}

