% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nouns_ING.R
\docType{data}
\name{nouns_ING}
\alias{nouns_ING}
\title{Zbior danych nouns_ING}
\format{lista wektorow napisow dlugosci 6977}
\usage{
data(nouns_ING)
}
\description{
Zbior zawiera rzeczowniki w formie podstawowej z kazdego wpisu z ramki wpisy_ING
}
\keyword{datasets}

