% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porównaj_wpisy.R
\name{porownaj_wpisy}
\alias{porownaj_wpisy}
\title{Funkcja wypisuje okresy czasowe dla ktorych wzmaga sie czestotliwosc
wystepowania slow ze slownika we wpsiach dla bankow}
\usage{
porownaj_wpisy(data_start = "2013-08-01", data_koniec = "2014-01-01",
  podzial = 10)
}
\arguments{
\item{data_start}{Poczatek rozpatrywanego okresu}

\item{data_koniec}{Koniec rozpatrywanego okresu}

\item{podzial}{liczba dni w analizowanym okresie}
}
\description{
Funkcja porownaj_wpisy() dla zadanego przedzialu czasowego analizuje czestotliwosci
wystepowania slow ze slownika we wpsiach dla bankow i w przypadku sie jej zwiekszenia
wypisuje przedzial czasowy w ktorym sie to znajduje oraz "podejrzane" slowa
}

