% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slownik_awarii_kart.R
\docType{data}
\name{slownik_awarii_kart}
\alias{slownik_awarii_kart}
\title{Zbior danych slownik_awarii_kart}
\format{wektor napisow dlugosci 12}
\usage{
data(slownik_awarii_kart)
}
\description{
Zbior zawiera wybrane rzeczowniki w formie podstawowej
dotyczace kart bankowych i bankomatow
}
\keyword{datasets}

