% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unikalne_ALIOR.R
\docType{data}
\name{unikalne_ALIOR}
\alias{unikalne_ALIOR}
\title{Zbior danych unikalne_ALIOR}
\format{lista wektorow napisow dlugosci 6920}
\usage{
data(unikalne_ALIOR)
}
\description{
Zbior zawiera rzeczowniki w formie podstawowej
z kazdego wpisu z ramki wpisy_ALIOR bez powtorzen
}
\keyword{datasets}

