% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unikalne_BZWBK.R
\docType{data}
\name{unikalne_BZWBK}
\alias{unikalne_BZWBK}
\title{Zbior danych unikalne_BZWBK}
\format{lista wektorow napisow dlugosci 25760}
\usage{
data(unikalne_BZWBK)
}
\description{
Zbior zawiera rzeczowniki w formie podstawowej
z kazdego wpisu z ramki wpisy_BZWBK bez powtorzen
}
\keyword{datasets}

