% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unikalne_ING.R
\docType{data}
\name{unikalne_ING}
\alias{unikalne_ING}
\title{Zbior danych unikalne_ING}
\format{lista wektorow napisow dlugosci 6977}
\usage{
data(unikalne_ING)
}
\description{
Zbior zawiera rzeczowniki w formie podstawowej
z kazdego wpisu z ramki wpisy_ING bez powtorzen
}
\keyword{datasets}

