% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpisy_ALIOR.R
\docType{data}
\name{wpisy_ALIOR}
\alias{wpisy_ALIOR}
\title{Zbior danych ALIOR}
\format{ramka danych z 6920 obserwacjami i 9 kolumnami}
\usage{
data(wpisy_ALIOR)
}
\description{
Zbior opisuje pelne dane opisujace wpisy z banku AliorBankSA
\itemize{
 \item id.  id postu
 \item parent_id.  id rodzica postu
 \item created_at. data utworzenia postu
 \item thread_id. id tematu
 \item subject. nazwa tematu (NA)
 \item user_name. nazwa tworcy postu
 \item user_link.  zrodlo tworcy postu
 \item source.  strona skad pochodzi wpis
 \item body.  tresc wpisu
}
}
\keyword{datasets}

