% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wybierz_wpisy.R
\name{wybierz_wpisy}
\alias{wybierz_wpisy}
\title{Funkcja wybiera wpisy zawierajace slowo w zadanym przedziale czasowym}
\usage{
wybierz_wpisy(nazwa_banku, slowo, data_start = "2013-09-01",
  data_koniec = "2014-01-01")
}
\arguments{
\item{nazwa_banku}{Nazwa banku, obecnie do wyboru "ALIOR", "ING", "BZWBK"}

\item{slowo}{Slowo ktore ma wystepowac we wpisie}

\item{data_start}{Poczatek rozpatrywanego okresu}

\item{data_koniec}{Koniec rozpatrywanego okresu}
}
\value{
wektor named_num czestotliwosci danego slowa
}
\description{
Funkcja wybierz_wpisy() dla zadanego banku i napisu zwraca wszystkie wpisy
w zadanym przedziale czasowym
}

